requirement.txt
openai-whisper
torch
transformers
datasets
pyaudio
mecab-python3
alkana
unidic-lite
TTS
sounddevice
boto3
google-api-python-client